Get the export from the /source folder instead of over the network. Check if the dat file exists before importing.

+++ fetch.sh
@@ -3,11 +3,14 @@ set -e
 
 mkdir -p /tmp/build
 
-curl --list-only https://dev-kartat.hsldev.com/poiminta/ | jq -r '.[0].name' > /tmp/allfiles.txt
+if [ ! -d "/source" ]; then
+    echo "Source not found. Add the /source dir and put a JORE export in it."
+    echo "Exiting!" 1>&2
+    exit 64
+fi
 
-grep -E '^.*\.zip$' /tmp/allfiles.txt > /tmp/latestfile.txt
+find /source -type f -name "*.zip" -exec basename {} \; | head -n 1 > /tmp/latestfile.txt
 export LATEST_FILE=`tail -1 /tmp/latestfile.txt`
-rm /tmp/allfiles.txt
 rm /tmp/latestfile.txt
 
 echo "Latest file is ${LATEST_FILE}"
@@ -21,7 +24,5 @@ fi
 
 rm -rf /tmp/build/*
 
-curl https://dev-kartat.hsldev.com/poiminta/${LATEST_FILE} --output /tmp/build/${LATEST_FILE}
-
-ln -s /tmp/build/${LATEST_FILE} /tmp/build/latest.zip
+ln -s /source/${LATEST_FILE} /tmp/build/latest.zip
 echo "Latest build can be accessed as /tmp/build/latest.zip"
