+++ fetch_local.sh
@@ -5,16 +5,26 @@ mkdir -p ./processed
 mkdir -p ./data
 mkdir -p ./tmp/build
 
-curl --list-only https://dev-kartat.hsldev.com/poiminta/ | jq -r '.[0].name' > ./tmp/allfiles.txt
+if [ ! -d "./source" ]; then
+    echo "Source not found. Add the ./source dir and put a JORE export in it."
+    echo "Exiting!" 1>&2
+    exit 64
+fi
 
-grep -E '^.*\.zip$' ./tmp/allfiles.txt > ./tmp/latestfile.txt
+find ./source -type f -name "*.zip" -exec basename {} \; | head -n 1 > ./tmp/latestfile.txt
 export LATEST_FILE=`tail -1 ./tmp/latestfile.txt`
-rm ./tmp/allfiles.txt
 rm ./tmp/latestfile.txt
 
-echo "Latest file is $LATEST_FILE"
+echo "Latest file is ${LATEST_FILE}"
 
-curl https://dev-kartat.hsldev.com/poiminta/${LATEST_FILE} --output /tmp/build/${LATEST_FILE}
+if [ -f "./tmp/build/$LATEST_FILE" ]
+then
+    echo "Latest build was already found, won't run it again"
+    echo "Exiting!" 1>&2
+    exit 64
+fi
 
-unzip ./tmp/build/$LATEST_FILE -d ./data
+rm -rf ./tmp/build/*
+
+unzip ./source/$LATEST_FILE -d ./data
 echo "Latest build unzipped into ./data"
